+++ scripts/generate.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-# coding=utf-8
+# -*- coding:utf-8 -*-
 
 import os
 import sys
@@ -12,7 +12,7 @@ def generate_file(version: str, nb_recurse: int):
     j2_env = Environment(loader=FileSystemLoader(os.path.dirname(os.path.abspath(__file__))),
                          trim_blocks=True)
     with open(os.path.dirname(os.path.realpath(__file__)) + '/../package.json', 'w') as f:
-        f.write(j2_env.get_template('package.json.jinja').render(version=version, nb_recurse=nb_recurse))
+        f.write(j2_env.get_template('package.jinja.json').render(version=version, nb_recurse=nb_recurse))
     # minimise json file
     with open(os.path.dirname(os.path.realpath(__file__)) + '/../package.json', 'r') as f:
         file_data = json.load(f)